+++ taxonomy-editor-frontend/src/backend-types/types.ts
@@ -31,3 +31,5 @@ type NodeType = {
 };
 
 export type RootEntriesAPIResponse = Array<NodeType[]>;
+
+export type SearchAPIResponse = string[];
