+++ taxonomy-editor-frontend/src/components/CreateNodeDialogContent.tsx
@@ -1,4 +1,5 @@
 import { useState } from "react";
+import ISO6391 from "iso-639-1";
 
 import { TextField, Button, FormControl, InputLabel } from "@mui/material";
 import DialogActions from "@mui/material/DialogActions";
@@ -6,10 +7,11 @@ import DialogContent from "@mui/material/DialogContent";
 import Select from "@mui/material/Select";
 import MenuItem from "@mui/material/MenuItem";
 import Alert from "@mui/material/Alert";
+import DialogTitle from "@mui/material/DialogTitle";
 
-import ISO6391 from "iso-639-1";
 
 import { createBaseURL } from "../../utils";
+
 
 type Props = {
   onCloseDialog: () => void;
@@ -18,7 +20,7 @@ type Props = {
   branchName: string;
 };
 
-const AddNodeDialogContent = ({
+const CreateNodeDialogContent = ({
   onCloseDialog,
   onSuccess,
   taxonomyName,
@@ -53,6 +55,7 @@ const AddNodeDialogContent = ({
 
   return (
     <>
+      <DialogTitle>Create new node</DialogTitle>
       <DialogContent>
         {isFailedSubmit && (
           <Alert severity="error">
@@ -103,4 +106,4 @@ const AddNodeDialogContent = ({
   );
 };
 
-export default AddNodeDialogContent;
+export default CreateNodeDialogContent;
